moved models into their respective files, deleted folders and fixed imports
