+++ README_en.md
@@ -181,8 +181,9 @@ model = build_segformer(config_handler)
 Initialization of Segformer from a configuration file.
 
 ```python
-from src.models.segformer.model import build_segformer
-config_handler = read_yaml_config(config_path) # обработчик конфигурационных файлов
+from src.models.segformer.segformer import build_segformer
+
+config_handler = read_yaml_config(config_path)  # обработчик конфигурационных файлов
 model = build_segformer(config_handler)
 ```
 
@@ -214,10 +215,11 @@ Training results of the model:
     <summary>  HRNet  </summary>
 
 Initialization of HRNet from a configuration file.
-    
+
 ```python
-from src.models.hrnet.model import build_hrnet
-config_handler = read_yaml_config(config_path) # Configuration file handler
+from src.models.hrnet.hrnet import build_hrnet
+
+config_handler = read_yaml_config(config_path)  # Configuration file handler
 model = build_hrnet(config_handler)
 ```
 
@@ -232,8 +234,9 @@ Training results of the model:
 Initialization of DeepLabV3 from a configuration file.
 
 ```python
-from src.models.deeplab.model import build_deeplab
-config_handler = read_yaml_config(config_path) # обработчик конфигурационных файлов
+from src.models.deeplab.deeplab import build_deeplab
+
+config_handler = read_yaml_config(config_path)  # обработчик конфигурационных файлов
 model = build_deeplab(config_handler)
 ```
