+++ src/infrastructure/models_inference/segformer/segformer_export.py
@@ -19,7 +19,7 @@ import sys
 import torch
 import transformers
 
-from src.models.segformer.model import SegFormer
+from src.models.segformer.segformer import SegFormer
 from src.models.inference import export_model_onnx, quantize_onnx
 from src.models.utils.config import read_yaml_config
