+++ src/infrastructure/models_tracking/deeplab_tracking.py
@@ -19,7 +19,7 @@ sys.path.append(f"../src/")
 
 from src.features.segmentation.dataset import get_train_dataset_by_config, get_val_dataset_by_config
 from src.features.segmentation.augmentation import get_augmentations_from_config, get_preprocess_from_config
-from src.models.deeplab.model import build_deeplab
+from src.models.deeplab import build_deeplab
 from src.infrastructure.tracking import tracking_experiment
 from src.models.utils.config import read_yaml_config
