+++ src/infrastructure/models_tracking/segformer_tracking.py
@@ -18,7 +18,7 @@ import sys
 sys.path.append(f"../src/")
 
 from src.features.segmentation.dataset import get_train_dataset_by_config, get_val_dataset_by_config
-from src.models.segformer.model import build_segformer
+from src.models.segformer import build_segformer
 from src.infrastructure.tracking import tracking_experiment
 from src.features.segmentation.augmentation import get_augmentations_from_config, get_preprocess_from_config
 from src.models.utils.config import read_yaml_config
